Refactor, simplify, add small error handling
I made some changes here to help you with the weird error udacities search method was hitting you with. It seems that if you search for things with certain letters it breaks. Like just type the letter 'o' and it things the search was empty and gives an error. But I checked and it wasn't sent as empty, so that isn't on you. Its on them.
Also, checkout that note on BEM.
And this is a great tool for Chrome while you are using dev tools for React. https://chrome.google.com/webstore/detail/react-developer-tools/fmkadmapgofadopljbjfkapdkoienihi?hl=en